move keyboard dismiss to new helper
